requirements.txt
matplotlib
numpy
Pillow
tqdm
wandb
torch==1.11.0
torchvision==0.12.0
numpy==1.26.3
tqdm
tensorboard
tensorboardX
ml-collections
scipy==1.11.3
timm
einops
torchmetrics
pytorch-ignite
monai
opencv-python